python/requirements.txt
bleak==0.21.1
keras==2.15.0
matplotlib==3.8.2
numpy==1.26.4
pyserial==3.5
